wormhole::commands::cli_commands

Enum Cli

pub enum Cli {
    Start(StatusPodArgs),
    Stop(StatusPodArgs),
    Template(TemplateArg),
    New(PodArgs),
    Inspect,
    GetHosts(GetHostsArgs),
    Tree(TreeArgs),
    Remove(RemoveArgs),
    Apply(PodConf),
    Restore(PodConf),
    Interrupt,
}

Variants

Start(StatusPodArgs)

Start the service

Stop(StatusPodArgs)

Stop the service

Template(TemplateArg)

Create a new network (template)

New(PodArgs)

Create a new pod and join a network if he have peers in arguments or create a new network

Inspect

Inspect a pod with its configuration, connections, etc

GetHosts(GetHostsArgs)

Get hosts for a specific file

Tree(TreeArgs)

Tree the folder structure from the given path and show hosts for each file

Remove(RemoveArgs)

Remove a pod from its network

Apply(PodConf)

Apply a new configuration to a pod

Restore(PodConf)

Restore many or a specifique file configuration

Interrupt

Stops the service