wormhole::config::types
Trait Config

pub trait Config: Sized {
    // Required methods
    fn write<S: AsRef<Path>>(&self, path: S) -> Result<(), Box<dyn Error>>;
    fn read<S: AsRef<Path>>(path: S) -> Result<Self, Box<dyn Error>>;
}

NOTE To add elements in the configuration file : To create a superior field like [field], create a new structure and add it to the Metadata struct Minors fields are named in the structure you added to Metadata the section name is the same as the name of the value of your new struct in Metadata

Required Methods

fn write<S: AsRef<Path>>(&self, path: S) -> Result<(), Box<dyn Error>>

fn read<S: AsRef<Path>>(path: S) -> Result<Self, Box<dyn Error>>

Object Safety

This trait is not object safe.