wormhole::ipc

Module answers

Structs

InspectInfo

PeerInfoIPC

Not to be confused with PeerInfo though the two are the same data, this one is exclusively the CLI messaging the other is exclusively for Network messaging this distinction is because of typing

StatusSuccess

Enums

ApplyConfigAnswer

CheckConfigAnswer

ConfigFileError

FreezeAnswer

GenerateConfigAnswer

GetHostsAnswer

InspectAnswer

ListPodsAnswer

NewAnswer

PodCreationError

RedundancyStatusAnswer

RemoveAnswer

RestartAnswer

ShowConfigAnswer

StatsPerFiletypeAnswer

StatusAnswer

TreeAnswer

UnfreezeAnswer