wormhole::pods::arbo

Struct Arbo

pub struct Arbo { /* private fields */ }

Implementations

impl Arbo

pub fn first_ino() -> InodeId

pub fn new() -> Self

pub fn overwrite_self(&mut self, entries: ArboIndex)

pub fn get_raw_entries(&self) -> ArboIndex

pub fn iter(&self) -> Iter<'_, InodeId, Inode>

pub fn inodes_mut(&mut self) -> ValuesMut<'_, InodeId, Inode>

pub fn get_special(name: &str, parent_ino: u64) -> Option<u64>

pub fn is_special(ino: u64) -> bool

pub fn is_local_only(ino: u64) -> bool

pub fn read_lock<'a>( arbo: &'a Arc<RwLock<Arbo>>, called_from: &'a str, ) -> Result<RwLockReadGuard<'a, Arbo>>

pub fn n_read_lock<'a>( arbo: &'a Arc<RwLock<Arbo>>, called_from: &'a str, ) -> WhResult<RwLockReadGuard<'a, Arbo>>

pub fn write_lock<'a>( arbo: &'a Arc<RwLock<Arbo>>, called_from: &'a str, ) -> Result<RwLockWriteGuard<'a, Arbo>>

pub fn n_write_lock<'a>( arbo: &'a Arc<RwLock<Arbo>>, called_from: &'a str, ) -> WhResult<RwLockWriteGuard<'a, Arbo>>

pub fn files_hosted_only_by<'a>( &'a self, host: &'a Address, ) -> impl Iterator<Item = &'a Inode> + use<'a>

pub fn add_inode(&mut self, inode: Inode) -> Result<(), MakeInodeError>

Insert a given Inode inside the local arbo

pub fn add_inode_from_parameters( &mut self, name: String, id: InodeId, parent_ino: InodeId, entry: FsEntry, perm: u16, ) -> Result<(), MakeInodeError>

Create a new Inode from the given parameters and insert it inside the local arbo

pub fn remove_children(&mut self, parent: InodeId, child: InodeId) -> Result<()>

pub fn n_remove_child( &mut self, parent: InodeId, child: InodeId, ) -> WhResult<()>

pub fn add_children(&mut self, parent: InodeId, child: InodeId) -> Result<()>

๐Deprecated

pub fn n_add_child(&mut self, parent: InodeId, child: InodeId) -> WhResult<()>

pub fn remove_inode(&mut self, id: InodeId) -> Result<Inode>

pub fn n_remove_inode(&mut self, id: InodeId) -> Result<Inode, RemoveInodeError>

Remove inode from the Arbo

pub fn get_inode(&self, ino: InodeId) -> Result<&Inode>

pub fn n_get_inode(&self, ino: InodeId) -> WhResult<&Inode>

pub fn mv_inode( &mut self, parent: InodeId, new_parent: InodeId, name: &String, new_name: &String, ) -> Result<()>

๐Deprecated

pub fn n_mv_inode( &mut self, parent: InodeId, new_parent: InodeId, name: &String, new_name: &String, ) -> WhResult<()>

pub fn n_get_inode_mut(&mut self, ino: InodeId) -> WhResult<&mut Inode>

pub fn get_path_from_inode_id(&self, inode_index: InodeId) -> Result<WhPath>

pub fn n_get_path_from_inode_id(&self, inode_index: InodeId) -> WhResult<WhPath>

Recursively traverse the Arbo tree from the Inode to form a path

Possible Errors: InodeNotFound: if the inode isnโt inside the tree

pub fn get_inode_child_by_name( &self, parent: &Inode, name: &String, ) -> Result<&Inode>

pub fn n_get_inode_child_by_name( &self, parent: &Inode, name: &String, ) -> WhResult<&Inode>

pub fn get_inode_from_path(&self, path: &WhPath) -> Result<&Inode>

pub fn set_inode_hosts( &mut self, ino: InodeId, hosts: Vec<Address>, ) -> Result<()>

pub fn n_set_inode_hosts( &mut self, ino: InodeId, hosts: Vec<Address>, ) -> WhResult<()>

pub fn add_inode_hosts( &mut self, ino: InodeId, new_hosts: Vec<Address>, ) -> Result<()>

Add hosts to an inode

Only works on inodes pointing files (no folders) Ignore already existing hosts to avoid duplicates

pub fn n_add_inode_hosts( &mut self, ino: InodeId, new_hosts: Vec<Address>, ) -> WhResult<()>

Add hosts to an inode

Only works on inodes pointing files (no folders) Ignore already existing hosts to avoid duplicates

pub fn remove_inode_hosts( &mut self, ino: InodeId, remove_hosts: Vec<Address>, ) -> Result<()>

Remove hosts from an inode

Only works on inodes pointing files (no folders)

pub fn set_inode_meta(&mut self, ino: InodeId, meta: Metadata) -> Result<()>

pub fn n_set_inode_meta(&mut self, ino: InodeId, meta: Metadata) -> WhResult<()>

pub fn set_inode_size(&mut self, ino: InodeId, size: u64) -> WhResult<()>

pub fn set_inode_xattr( &mut self, ino: InodeId, key: String, data: Vec<u8>, ) -> WhResult<()>

pub fn remove_inode_xattr(&mut self, ino: InodeId, key: String) -> WhResult<()>