wormhole::pods::filesystem::fs_interface

Struct FsInterface

pub struct FsInterface {
    pub network_interface: Arc<NetworkInterface>,
    pub disk: Box<dyn DiskManager>,
    pub file_handles: Arc<RwLock<FileHandleManager>>,
    pub mountpoint: PathBuf,
}

Fields
network_interface: Arc<NetworkInterface>disk: Box<dyn DiskManager>file_handles: Arc<RwLock<FileHandleManager>>mountpoint: PathBuf
Implementations

impl FsInterface

pub fn acknowledge_metadata( &self, ino: Ino, meta: Metadata, ) -> Result<Response, AcknoledgeSetAttrError>

pub fn setattr( &self, ino: Ino, mode: Option<u32>, uid: Option<u32>, gid: Option<u32>, size: Option<u64>, atime: Option<SystemTime>, mtime: Option<SystemTime>, ctime: Option<SystemTime>, file_handle: Option<UUID>, flags: Option<u32>, ) -> Result<Metadata, SetAttrError>

impl FsInterface

pub fn flush( &self, ino: Ino, handle: Option<&mut FileHandle>, ) -> Result<(), FlushError>

Flush file changes to all connected peers With a handle, the changed will be first attempted with the handle’s base signature the handle’s signature will be flushed. The dirty flag will be reset Without a handle, all peers will get a ‘FileChanged’ message and tracking peers will request the delta with their own base signature Non-tracking peers will always only get a ‘FileChanged’ message

pub fn accept_delta( &self, ino: Ino, meta: Metadata, sig: Signature, delta: Delta, ) -> Result<Response, FlushError>

Apply a delta received from the network deltas are in reference to a base signature, in case of signature mismatch [Request::DeltaRequest] is emitted back to get the correct diff

pub fn accept_file_changed( &self, ino: Ino, meta: Metadata, ) -> Result<Response, FlushError>

Acknowledge a file change and request the change contents if we are tracking the file

pub fn respond_delta( &self, ino: Ino, sig: Signature, origin: PeerId, ) -> Result<(), FlushError>

impl FsInterface

Provides functions to allow primitive handlers like Fuse & WinFSP to interract with wormhole

pub fn new( network_interface: Arc<NetworkInterface>, disk_manager: Box<dyn DiskManager>, mountpoint: PathBuf, ) -> Self

pub fn get_entry_from_name( &self, parent: Ino, name: &str, ) -> WhResult<Option<Inode>>

get an entry return Ok(None) if no permissions to access entries

pub fn get_inode_attributes(&self, ino: Ino) -> WhResult<Metadata>

pub fn recept_inode(&self, inode: Inode) -> Result<Response, MakeInodeError>

pub fn recept_redundancy( &self, ino: Ino, binary: Arc<Vec<u8>>, ) -> WhResult<Response>

pub fn recept_add_hosts(&self, id: Ino, hosts: &[PeerId]) -> WhResult<Response>

pub fn recept_remove_hosts( &self, id: Ino, hosts: &[PeerId], ) -> WhResult<Response>

pub fn send_file(&self, inode: Ino) -> Result<Response>

pub fn read_local_file(&self, inode: Ino) -> WhResult<Vec<u8>>

pub fn get_size_info(&self) -> Result<DiskSizeInfo>

Get complete filesystem size information including inode counts

impl FsInterface

pub fn create( &self, parent_ino: Ino, name: InodeName, entry: FsEntry, flags: OpenFlags, access: AccessMode, permissions: u16, ) -> Result<(Inode, UUID), CreateError>

pub fn make_inode( &self, parent_ino: Ino, name: InodeName, permissions: u16, entry: FsEntry, ) -> Result<Inode, MakeInodeError>

Create a new empty Inode, define its informations and register both in the network and in the local filesystem Immediately replicated to other peers

impl FsInterface

pub fn open( &self, ino: Ino, flags: OpenFlags, access: AccessMode, ) -> Result<UUID, OpenError>

impl FsInterface

pub fn get_file_data_sync( &self, ino: Ino, offset: usize, buf: &mut [u8], ) -> Result<usize, ReadError>

Panics

This function panics if called within an asynchronous execution context.

pub fn pull_file_sync(&self, ino: Ino) -> Result<Option<Vec<u8>>, PullError>

Pull the file from the network Affects the itree and the disk

Panics

This function panics if called within an asynchronous execution context.

pub fn get_whole_file_sync(&self, ino: Ino) -> Result<File, ReadError>

Get or pull the file from storage or network

Panics

This function panics if called within an asynchronous execution context.

pub fn get_local_file(&self, ino: Ino) -> Result<Option<File>, ReadError>

Get locally stored file as-is if it exists without accessing the network returns Ok(Some(file)) if the file is tracked returns Ok(None) if the file is not tracked

pub fn read_file( &self, file: Ino, offset: usize, buf: &mut [u8], file_handle: UUID, ) -> Result<usize, ReadError>

impl FsInterface

pub fn read_dir( &self, ino: Ino, ) -> Result<Vec<(u64, String, Metadata)>, ReadDirError>

pub fn readlink(&self, ino: Ino) -> Result<EntrySymlink, ReadLinkError>

impl FsInterface

pub fn release(&self, file_handle: UUID) -> Result<(), FlushError>

impl FsInterface

pub fn fuse_remove_inode( &self, parent: Ino, name: InodeName, ) -> Result<(), RemoveFileError>

pub fn remove_inode_locally(&self, id: Ino) -> Result<(), RemoveFileError>

pub fn remove_inode(&self, id: Ino) -> Result<(), RemoveFileError>

Remove an [Inode] from the ITree Immediately replicated to other peers

pub fn recept_remove_inode(&self, id: Ino) -> Result<Response, RemoveFileError>

impl FsInterface

pub fn set_meta_size(&self, ino: Ino, meta: Metadata) -> Result<(), RenameError>

pub fn rename( &self, parent: Ino, new_parent: Ino, name: InodeName, new_name: InodeName, overwrite: bool, ) -> Result<(), RenameError>

Rename a file, by changing its name but usually not its ino

overwrite: silently delete a file with the destination name

special: when using special name, completely switch behavior:

if overwrite, delete destination officialy

create a new destination inode

copy/move contents

delete the source inode

Immediately replicated to other peers

pub fn recept_rename( &self, parent: Ino, new_parent: Ino, name: InodeName, new_name: InodeName, overwrite: bool, ) -> Result<Response, RenameError>

impl FsInterface

pub fn write( &self, id: Ino, data: &[u8], offset: usize, file_handle: UUID, ) -> Result<usize, WriteError>

modifies the local file on disk marks the file handle as dirty, but does not immediately send the change to other peers

impl FsInterface

pub fn get_inode_xattr( &self, ino: Ino, key: &str, ) -> Result<Vec<u8>, GetXAttrError>

pub fn xattr_exists(&self, ino: Ino, key: &str) -> WhResult<bool>

pub fn list_inode_xattr(&self, ino: Ino) -> WhResult<Vec<String>>