wormhole::pods::network::network_interface

Struct NetworkInterface

pub struct NetworkInterface {
    pub arbo: Arc<RwLock<Arbo>>,
    pub mount_point: WhPath,
    pub to_network_message_tx: UnboundedSender<ToNetworkMessage>,
    pub to_redundancy_tx: UnboundedSender<RedundancyMessage>,
    pub next_inode: Mutex<InodeId>,
    pub callbacks: Callbacks,
    pub peers: Arc<RwLock<Vec<PeerIPC>>>,
    pub local_config: Arc<RwLock<LocalConfig>>,
    pub global_config: Arc<RwLock<GlobalConfig>>,
}

Fields
arbo: Arc<RwLock<Arbo>>mount_point: WhPathto_network_message_tx: UnboundedSender<ToNetworkMessage>to_redundancy_tx: UnboundedSender<RedundancyMessage>next_inode: Mutex<InodeId>callbacks: Callbackspeers: Arc<RwLock<Vec<PeerIPC>>>local_config: Arc<RwLock<LocalConfig>>global_config: Arc<RwLock<GlobalConfig>>
Implementations

impl NetworkInterface

pub fn new( arbo: Arc<RwLock<Arbo>>, mount_point: WhPath, to_network_message_tx: UnboundedSender<ToNetworkMessage>, to_redundancy_tx: UnboundedSender<RedundancyMessage>, next_inode: InodeId, peers: Arc<RwLock<Vec<PeerIPC>>>, local_config: Arc<RwLock<LocalConfig>>, global_config: Arc<RwLock<GlobalConfig>>, ) -> Self

pub fn get_next_inode(&self) -> Result<u64>

pub fn n_get_next_inode(&self) -> WhResult<u64>

TODO: Doc when reviews are finished

pub fn promote_next_inode(&self, new: u64) -> WhResult<()>

pub fn register_new_inode(&self, inode: Inode) -> Result<(), MakeInodeError>

Add the requested entry to the arbo and inform the network

pub fn n_rename( &self, parent: InodeId, new_parent: InodeId, name: &String, new_name: &String, overwrite: bool, ) -> Result<(), RenameError>

pub fn acknowledge_rename( &self, parent: InodeId, new_parent: InodeId, name: &String, new_name: &String, ) -> Result<(), RenameError>

pub fn acknowledge_new_file( &self, inode: Inode, _id: InodeId, ) -> Result<(), MakeInodeError>

Get a new inode, add the requested entry to the arbo and inform the network

pub fn unregister_inode(&self, id: InodeId) -> Result<(), RemoveInodeError>

Remove Inode from the Arbo and inform the network of the removal

pub fn acknowledge_unregister_inode( &self, id: InodeId, ) -> Result<Inode, RemoveInodeError>

Remove Inode from the Arbo

pub fn acknowledge_hosts_edition( &self, id: InodeId, hosts: Vec<Address>, ) -> WhResult<()>

pub fn send_file( &self, inode: InodeId, data: Vec<u8>, to: Address, ) -> Result<()>

pub fn write_file(&self, id: InodeId, new_size: usize) -> WhResult<()>

pub fn revoke_remote_hosts(&self, id: InodeId) -> WhResult<()>

pub fn add_inode_hosts(&self, ino: InodeId, hosts: Vec<Address>) -> WhResult<()>

pub fn update_hosts(&self, ino: InodeId, hosts: Vec<Address>) -> WhResult<()>

pub fn aknowledge_new_hosts( &self, id: InodeId, new_hosts: Vec<Address>, ) -> Result<()>

pub fn aknowledge_hosts_removal( &self, id: InodeId, new_hosts: Vec<Address>, ) -> Result<()>

pub fn update_metadata(&self, id: InodeId, meta: Metadata) -> WhResult<()>

pub fn apply_redundancy(&self, file_id: InodeId)

pub fn register_to_others(&self)

pub async fn request_arbo(&self, to: Address) -> Result<bool>

pub fn edit_peer_ip(&self, actual: Address, new: Address)

pub fn send_arbo(&self, to: Address, global_config_bytes: Vec<u8>) -> Result<()>

pub fn register_new_node(&self, socket: Address, addr: Address)

pub fn disconnect_peer(&self, addr: Address) -> Result<()>

pub async fn network_airport( network_reception: UnboundedReceiver<FromNetworkMessage>, fs_interface: Arc<FsInterface>, )

pub async fn contact_peers( peers_list: Arc<RwLock<Vec<PeerIPC>>>, rx: UnboundedReceiver<ToNetworkMessage>, )

pub async fn incoming_connections_watchdog( server: Arc<Server>, nfa_tx: UnboundedSender<FromNetworkMessage>, existing_peers: Arc<RwLock<Vec<PeerIPC>>>, )

impl NetworkInterface

pub fn pull_file_sync( &self, file: InodeId, ) -> Result<Option<Callback>, PullError>

impl NetworkInterface

pub async fn send_file_redundancy( &self, inode: InodeId, data: Arc<Vec<u8>>, to: Address, ) -> WhResult<Address>

impl NetworkInterface

pub fn set_inode_xattr( &self, ino: InodeId, key: String, data: Vec<u8>, ) -> WhResult<()>

pub fn recept_inode_xattr( &self, ino: InodeId, key: String, data: Vec<u8>, ) -> WhResult<()>

pub fn remove_inode_xattr(&self, ino: InodeId, key: String) -> WhResult<()>

pub fn recept_remove_inode_xattr( &self, ino: InodeId, key: String, ) -> WhResult<()>