wormhole::pods::whpath

Struct WhPath

pub struct WhPath {
    pub inner: String,
    pub kind: PathType,
}

Fields
inner: Stringkind: PathType
Implementations

impl WhPath

pub fn new() -> Self

pub fn push<T>(&mut self, segment: &T) -> &Self
where T: JoinPath + ?Sized,

Add a segment to the current WhPath. If the segment starts with a / or ./ or ../, the leading slash is removed. If the current WhPath is empty, the segment is added as is. If the current WhPath is not empty, the segment is added after adding a slash at the end of the current WhPath.

Examples

pub fn join<T>(&self, segment: &T) -> Self
where T: JoinPath + ?Sized,

Join the current path with a new segment. If the segment starts with a / or ./ or ../, the leading slash is removed. If the current path is empty, the segment is added as is. If the current path is not empty, the segment is added after adding a slash at the end of the current path. If the segment is hidden (is_hidden() return true), the segment is added as is, without adding a slash.

Examples

pub fn remove<T>(&mut self, delete_this_part: &T) -> &Self
where T: JoinPath + ?Sized,

Remove the requested part โ/my/file/path/โ.remove(โfile/pathโ) == โ/my/โ

pub fn rename<T>(&mut self, file_name: &T) -> &Self
where T: JoinPath + ?Sized,

Modify the path to match the new name. Can only modify the last element of the path.

pub fn kind(&self) -> PathType

pub fn update_kind(&mut self)

pub fn set_relative(self) -> Self

Change the path for โ./pathโ

pub fn set_absolute(self) -> Self

Change the path for โ/pathโ

pub fn remove_prefix(self) -> Self

Change the path for โpathโ

pub fn is_relative(&self) -> bool

pub fn is_absolute(&self) -> bool

pub fn has_no_prefix(&self) -> bool

pub fn is_empty(&self) -> bool

pub fn set_end(&mut self, end: bool) -> &Self

Put or not a โ/โ at the end

pub fn is_in<T>(&self, segment: &T) -> bool
where T: JoinPath + ?Sized,

Return true if the requested path is in the original path

pub fn get_end(&self) -> String

Give the last element of the path

pub fn get_folder(&self) -> String

Returns all but the last element

pub fn split_folder_file(&self) -> (String, String)

pub fn pop(&mut self) -> &Self

pub fn to_vector(&mut self) -> Vec<String>