wormhole::config::types

Trait Config

pub trait Config: Serialize + DeserializeOwned {
    // Provided methods
    fn write<S: AsRef<Path>>(&self, path: S) -> Result<(), Box<dyn Error>> { ... }
    fn read<S: AsRef<Path>>(path: S) -> Result<Self, Box<dyn Error>>
       where Self: Sized { ... }
    fn read_lock<'a, T: Config>(
        conf: &'a Arc<RwLock<T>>,
        called_from: &'a str,
    ) -> WhResult<RwLockReadGuard<'a, T>> { ... }
    fn write_lock<'a, T: Config>(
        conf: &'a Arc<RwLock<T>>,
        called_from: &'a str,
    ) -> WhResult<RwLockWriteGuard<'a, T>> { ... }
}

NOTE To add elements in the configuration file : To create a superior field like [field], create a new structure and add it to the Metadata struct Minors fields are named in the structure you added to Metadata the section name is the same as the name of the value of your new struct in Metadata

Provided Methods

fn write<S: AsRef<Path>>(&self, path: S) -> Result<(), Box<dyn Error>>

fn read<S: AsRef<Path>>(path: S) -> Result<Self, Box<dyn Error>>
where Self: Sized,

fn read_lock<'a, T: Config>( conf: &'a Arc<RwLock<T>>, called_from: &'a str, ) -> WhResult<RwLockReadGuard<'a, T>>

fn write_lock<'a, T: Config>( conf: &'a Arc<RwLock<T>>, called_from: &'a str, ) -> WhResult<RwLockWriteGuard<'a, T>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.