wormhole::pods::filesystem::fs_interface

Struct FsInterface

pub struct FsInterface {
    pub network_interface: Arc<NetworkInterface>,
    pub disk: Box<dyn DiskManager>,
    pub file_handles: Arc<RwLock<FileHandleManager>>,
    pub arbo: Arc<RwLock<Arbo>>,
}

Fields
network_interface: Arc<NetworkInterface>disk: Box<dyn DiskManager>file_handles: Arc<RwLock<FileHandleManager>>arbo: Arc<RwLock<Arbo>>
Implementations

impl FsInterface

pub fn acknowledge_metadata( &self, ino: InodeId, meta: Metadata, ) -> Result<(), AcknoledgeSetAttrError>

pub fn setattr( &self, ino: InodeId, mode: Option<u32>, uid: Option<u32>, gid: Option<u32>, size: Option<u64>, atime: Option<SystemTime>, mtime: Option<SystemTime>, ctime: Option<SystemTime>, file_handle: Option<UUID>, flags: Option<u32>, ) -> Result<Metadata, SetAttrError>

impl FsInterface

Provides functions to allow primitive handlers like Fuse & WinFSP to interract with wormhole

pub fn new( network_interface: Arc<NetworkInterface>, disk_manager: Box<dyn DiskManager>, arbo: Arc<RwLock<Arbo>>, ) -> Self

pub fn set_inode_meta(&self, ino: InodeId, meta: Metadata) -> Result<()>

๐Deprecated

pub fn get_entry_from_name( &self, parent: InodeId, name: String, ) -> Result<Inode>

pub fn get_inode_attributes(&self, ino: InodeId) -> Result<Metadata>

pub fn n_get_inode_attributes(&self, ino: InodeId) -> WhResult<Metadata>

pub fn read_dir(&self, ino: InodeId) -> Result<Vec<Inode>>

pub fn recept_inode(&self, inode: Inode) -> Result<(), MakeInodeError>

pub fn recept_redundancy( &self, id: InodeId, binary: Arc<Vec<u8>>, ) -> WhResult<()>

pub fn recept_binary(&self, id: InodeId, binary: Vec<u8>) -> Result<()>

pub fn recept_edit_hosts( &self, id: InodeId, hosts: Vec<Address>, ) -> WhResult<()>

pub fn recept_revoke_hosts( &self, id: InodeId, host: Address, meta: Metadata, ) -> Result<(), AcknoledgeSetAttrError>

pub fn recept_add_hosts(&self, id: InodeId, hosts: Vec<Address>) -> Result<()>

pub fn recept_remove_hosts( &self, id: InodeId, hosts: Vec<Address>, ) -> Result<()>

pub fn send_filesystem(&self, to: Address) -> Result<()>

pub fn register_new_node(&self, socket: Address, addr: Address)

pub fn send_file(&self, inode: InodeId, to: Address) -> Result<()>

pub fn read_local_file(&self, inode: InodeId) -> WhResult<Vec<u8>>

impl FsInterface

pub fn create( &self, parent_ino: u64, name: String, kind: SimpleFileType, flags: OpenFlags, access: AccessMode, permissions: u16, ) -> Result<(Inode, UUID), CreateError>

pub fn make_inode( &self, parent_ino: u64, name: String, permissions: u16, kind: SimpleFileType, ) -> Result<Inode, MakeInodeError>

Create a new empty Inode, define its informations and register both in the network and in the local filesystem

impl FsInterface

pub fn open( &self, ino: InodeId, flags: OpenFlags, access: AccessMode, ) -> Result<UUID, OpenError>

impl FsInterface

pub fn get_file_data( &self, file: InodeId, offset: usize, buf: &mut [u8], ) -> Result<usize, ReadError>

pub fn read_file( &self, file: InodeId, offset: usize, buf: &mut [u8], file_handle: UUID, ) -> Result<usize, ReadError>

impl FsInterface

pub fn release(&self, file_handle: UUID) -> WhResult<()>

impl FsInterface

pub fn fuse_remove_inode( &self, parent: InodeId, name: &OsStr, ) -> Result<(), RemoveFileError>

pub fn remove_inode_locally(&self, id: InodeId) -> Result<(), RemoveFileError>

pub fn remove_inode(&self, id: InodeId) -> Result<(), RemoveFileError>

pub fn recept_remove_inode(&self, id: InodeId) -> Result<(), RemoveFileError>

impl FsInterface

pub fn set_meta_size( &self, ino: InodeId, meta: Metadata, ) -> Result<(), RenameError>

pub fn rename( &self, parent: InodeId, new_parent: InodeId, name: &String, new_name: &String, overwrite: bool, ) -> Result<(), RenameError>

Rename a file, by changing its name but usually not its ino

overwrite: silently delete a file with the destination name

special: when using special name, completely switch behavior:

if overwrite, delete destination officialy

create a new destination inode

copy/move contents

delete the source inode

pub fn recept_rename( &self, parent: InodeId, new_parent: InodeId, name: &String, new_name: &String, overwrite: bool, ) -> Result<(), RenameError>

impl FsInterface

pub fn write( &self, id: InodeId, data: &[u8], offset: usize, file_handle: UUID, ) -> Result<usize, WriteError>

impl FsInterface

pub fn get_inode_xattr( &self, ino: InodeId, key: &String, ) -> Result<Vec<u8>, GetXAttrError>

pub fn xattr_exists(&self, ino: InodeId, key: &String) -> WhResult<bool>

pub fn list_inode_xattr(&self, ino: InodeId) -> WhResult<Vec<String>>